backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1
Pillow==10.1.0
python-dotenv==1.0.0
scipy==1.11.3
matplotlib==3.8.0
python-multipart==0.0.6